Rename particle.version.py to particle._version.py, to hide the file
